Eliminate a trait that we no longer need.
